+++ NEWS.md
@@ -4,6 +4,7 @@ broom 0.3.6.9000
 * Added tidiers for "plm" (panel linear model) objects from the plm package.
 * Added `tidy.coeftest` for coeftest objects from the lmtest package.
 * Set up `tidy.lm` to work with "mlm" (multiple linear model) objects (those with multiple response columns).
+* Added `tidy` and `glance` for "biglm" and "bigglm" objects from the biglm package.
 
 broom 0.3.6
 -----------